begin PIO conversion
