--- Helpers.cpp
@@ -1,58 +1,0 @@
-/*
-  Helpers.h - Helper functions to be used by OPAL FW on PJRC Teensy 4.x board
-
-  Part of OpenGalvo - OPAL Firmware
-
-  Copyright (c) 2020-2021 Daniel Olsson
-
-  OPAL Firmware is free software: you can redistribute it and/or modify
-  it under the terms of the GNU General Public License as published by
-  the Free Software Foundation, either version 3 of the License, or
-  (at your option) any later version.
-
-  OPAL Firmware is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU General Public License for more details.
-
-  You should have received a copy of the GNU General Public License
-  along with OPAL Firmware.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#include <Arduino.h>
-
-uint64_t nanos()  // Code by luni @ https://forum.pjrc.com/threads/60493-A-Higher-Resolution-Micros
-{
-    static uint32_t oldCycles = ARM_DWT_CYCCNT;
-    static uint64_t highValue = 0;
-
-    uint32_t newCycles = ARM_DWT_CYCCNT;
-    if (newCycles < oldCycles)
-    {
-        highValue += 0x0000'0001'0000'0000;
-    }
-    oldCycles = newCycles;
-    return (highValue | newCycles) * (1E9/F_CPU);
-}
-
-void printDouble( double val, unsigned int precision){
-// prints val with number of decimal places determine by precision
-// NOTE: precision is 1 followed by the number of zeros for the desired number of decimial places
-// example: printDouble( 3.1415, 100); // prints 3.14 (two decimal places)
-
-    Serial.print (int(val));  //prints the int part
-    Serial.print("."); // print the decimal point
-    unsigned int frac;
-    if(val >= 0)
-      frac = (val - int(val)) * precision;
-    else
-       frac = (int(val)- val ) * precision;
-    int frac1 = frac;
-    while( frac1 /= 10 )
-        precision /= 10;
-    precision /= 10;
-    while(  precision /= 10)
-        Serial.print("0");
-
-    Serial.println(frac,DEC) ;
-}
